[Refactor] Extract artifact path and reference artifact sync services
Move normalize_kb_relative_path / resolve_kb_sandbox_path into
datus/storage/artifact_path.py and the reference SQL / reference template
YAML-to-KB sync into per-package artifact_sync modules. The sync functions
treat the YAML artifact as the source of truth: re-syncing an existing
business ID upserts the stored row instead of skipping it, so the
build_mode-based skip is gone (GenerationHooks keeps the parameter as an
ignored compatibility shim until its removal).
