+++ tests/unit_tests/cli/test_generation_hooks.py
@@ -22,12 +22,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 from datus_db_core import connector_registry
 
-from datus.cli.generation_hooks import (
-    GenerationCancelledException,
-    GenerationHooks,
-    normalize_kb_relative_path,
-    resolve_kb_sandbox_path,
-)
+from datus.cli.generation_hooks import GenerationCancelledException, GenerationHooks
 from datus.tools.func_tool.base import FuncToolResult
 from datus.tools.func_tool.filesystem_tools import FilesystemFuncTool  # noqa: F401
 from datus.tools.func_tool.generation_evidence import GenerationEvidence
@@ -519,224 +514,6 @@ class TestSyncToStorage:
         ):
             result = await hooks._sync_to_storage("/tmp/file.yaml", "semantic")
         assert "**Sync error:** disk full" in result
-
-
-# ---------------------------------------------------------------------------
-# Tests: _sync_reference_sql_to_db / _sync_reference_template_to_db
-# ---------------------------------------------------------------------------
-
-
-class TestSyncReferenceSqlToDb:
-    def test_valid_template_yaml(self, tmp_path):
-        import yaml
-
-        from datus.cli.generation_hooks import GenerationHooks
-
-        yaml_file = tmp_path / "tpl.yaml"
-        yaml_file.write_text(
-            yaml.dump(
-                {
-                    "sql": "SELECT * FROM t WHERE x = '{{val}}'",
-                    "name": "test_reference_sql",
-                    "summary": "Test reference sql",
-                    "search_text": "test reference sql val",
-                    "subject_tree": "Sales/Revenue",
-                    "tags": "test",
-                }
-            )
-        )
-
-        mock_config = MagicMock()
-
-        with (
-            patch("datus.cli.generation_hooks.ReferenceSqlRAG") as mock_rag_cls,
-            patch(
-                "datus.storage.reference_sql.init_utils.exists_reference_sql",
-                return_value=set(),
-            ),
-            patch(
-                "datus.storage.reference_sql.init_utils.gen_reference_sql_id",
-                return_value="new_id",
-            ),
-        ):
-            mock_rag = mock_rag_cls.return_value
-            mock_rag.upsert_batch = MagicMock()
-
-            result = GenerationHooks._sync_reference_sql_to_db(str(yaml_file), mock_config)
-
-        assert result["success"] is True
-        assert "Synced" in result["message"]
-
-    def test_missing_sql_field(self, tmp_path):
-        import yaml
-
-        from datus.cli.generation_hooks import GenerationHooks
-
-        yaml_file = tmp_path / "bad.yaml"
-        yaml_file.write_text(yaml.dump({"name": "no_sql"}))
-
-        result = GenerationHooks._sync_reference_sql_to_db(str(yaml_file), MagicMock())
-        assert result["success"] is False
-        assert "No reference_sql data" in result["error"]
-
-    def test_duplicate_skipped(self, tmp_path):
-        import yaml
-
-        from datus.cli.generation_hooks import GenerationHooks
-
-        yaml_file = tmp_path / "dup.yaml"
-        yaml_file.write_text(yaml.dump({"sql": "SELECT 1", "name": "dup", "summary": "x", "search_text": "x"}))
-
-        mock_config = MagicMock()
-        with (
-            patch("datus.cli.generation_hooks.ReferenceSqlRAG"),
-            patch(
-                "datus.storage.reference_sql.init_utils.exists_reference_sql",
-                return_value={"existing_id"},
-            ),
-            patch(
-                "datus.storage.reference_sql.init_utils.gen_reference_sql_id",
-                return_value="existing_id",
-            ),
-        ):
-            result = GenerationHooks._sync_reference_sql_to_db(str(yaml_file), mock_config)
-
-        assert result["success"] is True
-        assert "already exists" in result["message"]
-
-
-class TestSyncReferenceTemplateToDb:
-    def test_valid_template_yaml(self, tmp_path):
-        import yaml
-
-        yaml_file = tmp_path / "tpl.yaml"
-        yaml_file.write_text(
-            yaml.dump(
-                {
-                    "sql": "SELECT * FROM t WHERE x = '{{val}}'",
-                    "name": "test_reference_template",
-                    "summary": "Test reference template",
-                    "search_text": "test reference template val",
-                    "subject_tree": "Sales/Revenue",
-                    "comment": "Helpful template",
-                    "tags": "test",
-                }
-            )
-        )
-
-        mock_config = MagicMock()
-
-        with (
-            patch("datus.storage.reference_template.store.ReferenceTemplateRAG") as mock_rag_cls,
-            patch(
-                "datus.storage.reference_template.init_utils.exists_reference_templates",
-                return_value=set(),
-            ),
-            patch(
-                "datus.storage.reference_template.init_utils.gen_reference_template_id",
-                return_value="new_tpl_id",
-            ),
-            patch(
-                "datus.storage.reference_template.template_file_processor.extract_template_parameters",
-                return_value=[{"name": "val", "type": "string"}],
-            ),
-        ):
-            mock_rag = mock_rag_cls.return_value
-            mock_rag.upsert_batch = MagicMock()
-
-            result = GenerationHooks._sync_reference_template_to_db(str(yaml_file), mock_config)
-
-        assert result["success"] is True
-        assert "Synced reference template" in result["message"]
-        mock_rag.upsert_batch.assert_called_once()
-        stored = mock_rag.upsert_batch.call_args.args[0][0]
-        assert stored == {
-            "id": "new_tpl_id",
-            "name": "test_reference_template",
-            "template": "SELECT * FROM t WHERE x = '{{val}}'",
-            "parameters": json.dumps([{"name": "val", "type": "string"}]),
-            "comment": "Helpful template",
-            "summary": "Test reference template",
-            "search_text": "test reference template val",
-            "filepath": str(yaml_file),
-            "subject_path": ["Sales", "Revenue"],
-            "tags": "test",
-        }
-
-    def test_missing_sql_field(self, tmp_path):
-        import yaml
-
-        yaml_file = tmp_path / "bad.yaml"
-        yaml_file.write_text(yaml.dump({"name": "no_sql"}))
-
-        result = GenerationHooks._sync_reference_template_to_db(str(yaml_file), MagicMock())
-
-        assert result["success"] is False
-        assert "No reference_template data" in result["error"]
-
-    def test_blank_sql_returns_error(self, tmp_path):
-        import yaml
-
-        yaml_file = tmp_path / "blank.yaml"
-        yaml_file.write_text(yaml.dump({"sql": "   ", "name": "blank"}))
-
-        result = GenerationHooks._sync_reference_template_to_db(str(yaml_file), MagicMock())
-
-        assert result["success"] is False
-        assert "non-empty string" in result["error"]
-
-    def test_duplicate_skipped(self, tmp_path):
-        import yaml
-
-        yaml_file = tmp_path / "dup.yaml"
-        yaml_file.write_text(yaml.dump({"sql": "SELECT 1", "name": "dup_tpl"}))
-
-        mock_config = MagicMock()
-        with (
-            patch("datus.storage.reference_template.store.ReferenceTemplateRAG"),
-            patch(
-                "datus.storage.reference_template.init_utils.exists_reference_templates",
-                return_value={"existing_tpl_id"},
-            ),
-            patch(
-                "datus.storage.reference_template.init_utils.gen_reference_template_id",
-                return_value="existing_tpl_id",
-            ),
-        ):
-            result = GenerationHooks._sync_reference_template_to_db(str(yaml_file), mock_config)
-
-        assert result["success"] is True
-        assert "already exists" in result["message"]
-
-    def test_storage_error_returns_failure(self, tmp_path):
-        import yaml
-
-        yaml_file = tmp_path / "boom.yaml"
-        yaml_file.write_text(yaml.dump({"sql": "SELECT 1", "name": "boom_tpl"}))
-
-        mock_config = MagicMock()
-        with (
-            patch("datus.storage.reference_template.store.ReferenceTemplateRAG") as mock_rag_cls,
-            patch(
-                "datus.storage.reference_template.init_utils.exists_reference_templates",
-                return_value=set(),
-            ),
-            patch(
-                "datus.storage.reference_template.init_utils.gen_reference_template_id",
-                return_value="boom_id",
-            ),
-            patch(
-                "datus.storage.reference_template.template_file_processor.extract_template_parameters",
-                return_value=[],
-            ),
-        ):
-            mock_rag = mock_rag_cls.return_value
-            mock_rag.upsert_batch.side_effect = RuntimeError("boom")
-
-            result = GenerationHooks._sync_reference_template_to_db(str(yaml_file), mock_config)
-
-        assert result["success"] is False
-        assert result["error"] == "boom"
 
 
 # ---------------------------------------------------------------------------
@@ -1383,49 +1160,6 @@ class TestResolvePathCommonpathValueError:
 
 
 # ---------------------------------------------------------------------------
-# Tests: normalize_kb_relative_path (pure function)
-# ---------------------------------------------------------------------------
-
-
-class TestNormalizeKbRelativePath:
-    def test_prepends_when_prefix_missing(self):
-        assert normalize_kb_relative_path("orders.yaml", "semantic") == "semantic_models/orders.yaml"
-
-    def test_prepends_for_sql_summary(self):
-        assert normalize_kb_relative_path("q_001.yaml", "sql_summary") == "sql_summaries/q_001.yaml"
-
-    def test_metric_kind_co_locates_with_semantic_models(self):
-        """metrics live under semantic_models/metrics/ — same root as semantic."""
-        assert (
-            normalize_kb_relative_path("metrics/orders_metrics.yaml", "metric")
-            == "semantic_models/metrics/orders_metrics.yaml"
-        )
-
-    def test_idempotent_when_prefix_already_correct(self):
-        already = "semantic_models/orders.yaml"
-        assert normalize_kb_relative_path(already, "semantic") == already
-
-    def test_passes_through_paths_in_other_kinds(self):
-        path = "sql_summaries/q_001.yaml"
-        assert normalize_kb_relative_path(path, "semantic") == path
-
-    def test_absolute_paths_unchanged(self):
-        assert normalize_kb_relative_path("/abs/path/orders.yaml", "semantic") == "/abs/path/orders.yaml"
-
-    def test_empty_path_unchanged(self):
-        assert normalize_kb_relative_path("", "semantic") == ""
-
-    def test_dot_path_unchanged(self):
-        assert normalize_kb_relative_path(".", "semantic") == "."
-
-    def test_parent_traversal_unchanged(self):
-        assert normalize_kb_relative_path("../../etc/passwd", "semantic") == "../../etc/passwd"
-
-    def test_unknown_kind_unchanged(self):
-        assert normalize_kb_relative_path("orders.yaml", "unknown") == "orders.yaml"
-
-
-# ---------------------------------------------------------------------------
 # Tests: hook + tool agreement — _resolve_path finds files written by the
 # tool (naked filename path was a normalizer concern; with normalizer gone
 # the LLM writes the full prefix, so the hook resolver must keep returning
@@ -1455,52 +1189,3 @@ class TestHookAndToolPathAgreement:
         on_disk = subject_root / "semantic_models" / "orders.yml"
         assert os.path.realpath(resolved) == os.path.realpath(str(on_disk))
         assert Path(resolved).is_file()
-
-
-# ---------------------------------------------------------------------------
-# Tests: resolve_kb_sandbox_path — used by workflow-mode _save_to_db() helpers
-# ---------------------------------------------------------------------------
-
-
-class TestResolveKbSandboxPath:
-    def test_empty_path_returns_none(self, tmp_path):
-        assert resolve_kb_sandbox_path("", "sql_summary", str(tmp_path)) is None
-
-    def test_bare_filename_is_prefixed_under_sandbox(self, tmp_path):
-        kb = tmp_path
-        resolved = resolve_kb_sandbox_path("q_001.yaml", "sql_summary", str(kb))
-        assert resolved == os.path.normpath(str(kb / "sql_summaries" / "q_001.yaml"))
-
-    def test_fully_prefixed_relative_path_passes_through(self, tmp_path):
-        kb = tmp_path
-        resolved = resolve_kb_sandbox_path("sql_summaries/q.yaml", "sql_summary", str(kb))
-        assert resolved == os.path.normpath(str(kb / "sql_summaries" / "q.yaml"))
-
-    def test_absolute_path_inside_sandbox_accepted(self, tmp_path):
-        kb = tmp_path
-        (kb / "sql_summaries").mkdir(parents=True)
-        inside = kb / "sql_summaries" / "q.yaml"
-        inside.write_text("x")
-        resolved = resolve_kb_sandbox_path(str(inside), "sql_summary", str(kb))
-        assert os.path.realpath(resolved) == os.path.realpath(str(inside))
-
-    def test_absolute_path_outside_sandbox_rejected(self, tmp_path):
-        """A fabricated absolute path outside the sandbox must be refused so
-        _save_to_db never syncs an arbitrary on-disk file."""
-        assert resolve_kb_sandbox_path("/etc/passwd", "sql_summary", str(tmp_path)) is None
-
-    def test_traversal_escape_rejected(self, tmp_path):
-        """``../../etc/passwd`` resolves outside the sandbox → rejected."""
-        assert resolve_kb_sandbox_path("../../etc/passwd", "sql_summary", str(tmp_path)) is None
-
-    def test_unknown_kind_no_containment_check(self, tmp_path):
-        """For an unknown kind we cannot compute a sandbox — fall back to
-        just normalizing against knowledge_base_dir."""
-        resolved = resolve_kb_sandbox_path("foo.yaml", "unknown", str(tmp_path))
-        assert resolved == os.path.normpath(str(tmp_path / "foo.yaml"))
-
-    def test_commonpath_value_error_fails_closed(self, tmp_path):
-        """Simulate os.path.commonpath raising (e.g. mixed drives on
-        Windows) — the resolver must fail closed with None."""
-        with patch("datus.cli.generation_hooks.os.path.commonpath", side_effect=ValueError("mixed drives")):
-            assert resolve_kb_sandbox_path("q.yaml", "sql_summary", str(tmp_path)) is None
